wrangler/requirements.txt
Twisted==13.2.0
ago==0.0.5
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.8
celery==3.1.5
docker-py==0.1.5
kombu==3.0.6
mock==1.0.1
nose==1.3.0
python-nginx==0.1
pytz==2013.8
requests==2.0.1
shipper==0.1a
six==1.4.1
texttable==0.8.1
treq==0.2.0
wsgiref==0.1.2
zope.interface==4.0.5
